README.mdの更新、テストデータ投入コマンドを作成、seederの整理
