+++ src/database/seeders/DatabaseSeeder.php
@@ -2,9 +2,6 @@
 
 namespace Database\Seeders;
 
-use App\Models\User;
-use App\Models\Blog;
-use App\Models\Tag;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
@@ -15,26 +12,5 @@ class DatabaseSeeder extends Seeder
     /**
      * Seed the application's database.
      */
-    public function run(): void
-    {
-        // 固定テストユーザ作成
-        User::factory()->create([
-            'user_id' => 'test',
-            'email' => 'test@example.com',
-            'password' => 'test',
-        ]);
-        
-        // 3人のユーザそれぞれに5件のブログを作成
-        User::factory()
-            ->count(3)
-            ->has(
-                Blog::factory()
-                    ->count(10)
-                    ->hasAttached(
-                        Tag::factory()->count(rand(0,5))
-                    )
-            )
-            ->create();
-
-    }
+    public function run(): void {}
 }
